// src/lib/const/routes.enum.ts
export enum routes {
  add = 'add',
  edit = 'edit',
  view = 'view',
  delete = 'delete',
  leads = 'leads',
  messageTemplate = 'message-template',
  courses = 'courses',
  categories = 'categories',
  campaigns = 'campaigns',
  messages = 'messages',
  users = 'users',
  whatsapp = 'whatsapp',
  labs = 'labs',
  lecturers = 'lecturers',
  employees = 'employees',
  courseGroups = 'course-groups',
  groupsSchedule = 'groups-schedule',
  groupsReport = 'groups-report',
  lectures = 'lectures',
  students = 'students',
  branches = 'branches',
  cities = 'cities',
  areas = 'areas',
  agreements = 'agreements',
  qualificationDescription = 'qualification-description',
  qualificationIssuer = 'qualification-issuer',
  qualificationType = 'qualification-type',
  receipts = 'receipts',
  expenses = 'expenses',
  inventory = 'inventory',
  profitLoss = 'profit-loss',
  absences = 'absences',
  roles = 'roles',
  collectionAlerts = 'collection-alerts',
}