// src/app/[locale]/lecturers/(components)/lecturer.interface.ts
export interface ILecturer {
  id: number
  name: string
  email: string
  emailVerified: string
  image: string
  phone: string
  address: string
  nationalId: string
  cityId: number
  education: string
  coursesIds: string[]
  userId: number
  rank: string
  salary: number
  salaryTypeId: number
  specialization: string
  salaryTypeName: string
}